tests/wpt/web-platform-tests/fullscreen/api/element-request-fullscreen-not-allowed.html

<!DOCTYPE html>
<title>Element#requestFullscreen() when not allowed to request fullscreen</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(t => {
  const div = document.querySelector("div");

  document.addEventListener("fullscreenerror", t.step_func_done(event => {
    assert_equals(event.target, document, "event.target");
    assert_false(event.bubbles, "event.bubbles");
    assert_false(event.cancelable, "event.cancelable");
  }));

  div.requestFullscreen();
});
</script>
